helper-apps/cortex-autogen/requirements.txt
azure-storage-queue
azure-functions
azure-search-documents
pyautogen==0.3.0
redis
pymongo
requests
azure-storage-blob
mysql-connector-python